create schema module/ create channel db query
